NodeTasksHandler -> NodeAlgorithmTasksHandler
