+++ mipengine/controller/algorithm_executor_nodes.py
@@ -6,11 +6,12 @@ from typing import List
 from typing import Optional
 from typing import Tuple
 
+from mipengine.controller.algorithm_execution_tasks_handler import (
+    INodeAlgorithmTasksHandler,
+)
 from mipengine.controller.algorithm_executor_node_data_objects import NodeData
 from mipengine.controller.algorithm_executor_node_data_objects import SMPCTableNames
 from mipengine.controller.algorithm_executor_node_data_objects import TableName
-from mipengine.controller.node_tasks_handler_interface import INodeTasksHandler
-from mipengine.controller.node_tasks_handler_interface import IQueuedUDFAsyncResult
 from mipengine.node_tasks_DTOs import NodeSMPCDTO
 from mipengine.node_tasks_DTOs import NodeTableDTO
 from mipengine.node_tasks_DTOs import NodeUDFDTO
@@ -18,6 +19,11 @@ from mipengine.node_tasks_DTOs import TableData
 from mipengine.node_tasks_DTOs import TableSchema
 from mipengine.node_tasks_DTOs import UDFKeyArguments
 from mipengine.node_tasks_DTOs import UDFPosArguments
+
+
+class AsyncResult:
+    def get(self, timeout=None):
+        pass
 
 
 class _INode(ABC):
@@ -66,13 +72,11 @@ class _INode(ABC):
         func_name: str,
         positional_args: UDFPosArguments,
         keyword_args: UDFKeyArguments,
-    ) -> IQueuedUDFAsyncResult:
+    ) -> AsyncResult:
         pass
 
     @abstractmethod
-    def get_queued_udf_result(
-        self, async_result: IQueuedUDFAsyncResult
-    ) -> List[TableName]:
+    def get_queued_udf_result(self, async_result: AsyncResult) -> List[TableName]:
         pass
 
     @abstractmethod
@@ -85,9 +89,9 @@ class _Node(_INode, ABC):
         self,
         request_id: str,
         context_id: str,
-        node_tasks_handler: INodeTasksHandler,
+        node_tasks_handler: INodeAlgorithmTasksHandler,
     ):
-        self._node_tasks_handler: INodeTasksHandler = node_tasks_handler
+        self._node_tasks_handler: INodeAlgorithmTasksHandler = node_tasks_handler
         self.node_id: str = self._node_tasks_handler.node_id
         self.request_id: str = request_id
         self.context_id: str = context_id
@@ -182,7 +186,7 @@ class _Node(_INode, ABC):
         positional_args: UDFPosArguments,
         keyword_args: UDFKeyArguments,
         use_smpc: bool = False,
-    ) -> IQueuedUDFAsyncResult:
+    ) -> AsyncResult:
         return self._node_tasks_handler.queue_run_udf(
             request_id=self.request_id,
             context_id=self.context_id,
@@ -264,9 +268,7 @@ class LocalNode(_Node):
         )
         return [TableName(view) for view in views]
 
-    def get_queued_udf_result(
-        self, async_result: IQueuedUDFAsyncResult
-    ) -> List[NodeData]:
+    def get_queued_udf_result(self, async_result: AsyncResult) -> List[NodeData]:
         node_udf_results = self._node_tasks_handler.get_queued_udf_result(async_result)
         udf_results = []
         for result in node_udf_results.results:
@@ -308,9 +310,7 @@ def create_node_table_from_node_table_dto(node_table_dto: NodeTableDTO):
 
 
 class GlobalNode(_Node):
-    def get_queued_udf_result(
-        self, async_result: IQueuedUDFAsyncResult
-    ) -> List[TableName]:
+    def get_queued_udf_result(self, async_result: AsyncResult) -> List[TableName]:
         node_udf_results = self._node_tasks_handler.get_queued_udf_result(async_result)
         results = []
         for result in node_udf_results.results:
